pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "website-security-scanner"
version = "2.0.0"
description = "Low-code platform security scanner for comparative security analysis."
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Bachelor Thesis Project" }]
license = { text = "UNLICENSED" }
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "pyOpenSSL>=22.0.0",
    "cryptography>=3.4.8",
    "urllib3>=1.26.0",
    "html5lib>=1.1",
    "tqdm>=4.64.0",
    "jinja2>=3.1.0",
    "python-nmap>=0.7.1",
    "dnspython>=2.2.0",
    "httpx>=0.23.0",
    "openpyxl>=3.0.10",
    "reportlab>=3.6.0",
    "pyyaml>=6.0",
    "colorlog>=6.7.0",
    "netaddr>=0.8.0",
    "builtwith>=1.3.4",
    "python-whois>=0.7.3",
    "colorama>=0.4.5",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "python-socketio>=5.9.0",
    "eventlet>=0.33.0",
    "python-engineio>=4.7.0",
    "schedule>=1.2.0",
]

[project.scripts]
wss = "website_security_scanner.cli.cli:main"
wss-web = "website_security_scanner.web.run_server:main"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"website_security_scanner.web" = [
    "templates/*.html",
    "static/**/*",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]
